Improve rake task, devide poly_work.rb
